refactor: 계산 기능 추상화
- CalculationFunction 클래스를 통해 계산 기능을 추상화했습니다.
- AdditionFunction 클래스를 통해 덧셈 기능을 구현하였습니다.
- 뺄셈 등 다른 연산 기능이 쉽게 확장될 수 있도록 구조화했습니다.
