+++ src/calculator/calculation-function/calculation-function.js
@@ -1,0 +1,15 @@
+export class CalculationFunction {
+  #result = 0;
+
+  calculateString() {
+    throw new Error('calculateString()은 반드시 구현되어야 합니다.');
+  }
+
+  getResult() {
+    return this.#result;
+  }
+
+  addToResult(value) {
+    this.#result += value;
+  }
+}
